requirements.txt
fsspec
ffmpeg-python
praat-parselmouth
pyworld
matplotlib
soundfile
scikit-learn
scipy
torchcrepe
tensorboard
transformers
tqdm
librosa
omegaconf
torch
jax
flax
optax
orbax-checkpoint
einops
fairseq
flask
flask_cors
gin
gin_config
local_attention
numpy
PyYAML
resampy
scikit_learn
wave
pysimplegui
sounddevice
gradio
faiss-cpu
